pyproject.toml
[tool.poetry]
name = "martinuvsekac-cz"
version = "0.1.0"
description = ""
authors = ["Martin Rosenberg <nitram.rosenberg@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {version = "^0.95.0", extras = ["all"]}
jinja2 = "^3.1.2"
tortoise-orm = {version = "^0.19.3", extras = ["asyncpg"]}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
fastapi-sessions = "^0.3.2"
sqlalchemy = "^2.0.9"
psycopg2 = "^2.9.6"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
aerich = "^0.7.1"


[tool.aerich]
tortoise_orm = "db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
